requirements.txt
pytorch==1.12.1
einops==0.4.0
higher==0.2.1
hydra-core==1.2.0
transformers==4.23.1
datasets==1.18.3
matplotlib==3.6.1
spacy==3.4.1
scipy==1.9.2
scikit-learn==1.0.2
nltk==3.7
jupyter==1.0.0